Add option install a different image at runtime
Add a new behaviour, at install time, to use a different runtime image.
In this case, the install behaviour is slightly changed. It does
download the runtime image the user will need locally, but will fetch
the package from the registry instead of pulling the image if not
available locally. This avoids the user to suffer 2 downloads of a
similar image, and take non-required space on his drive.

This behaviour is introduced thanks to the inspector package provideing a single interface to inspect an image name.
Currently, inspection is supported through:
- inspecting local image using the docker daemon
- fetching the metadata from the image registry

Always fetching the package details will also allow to just update
packages without having to pull images, eventually.

Solves part of #94